Consolidated test fxns, adjusted fxn names, added new accessor fxns
